using setTargetResolution with normalized values on VideoCapture use case
